12352: **IllegalArgumentException in org.labkey.filecontent.FileContentServiceImpl.isFileRootDisabled() 12261: **IllegalStateException from org.labkey.api.module.SafeFlushResponseWrapper.getOutputStream() 12175: **NullPointerException from org.labkey.query.reports.ReportServiceImpl.createReportInstance() 12081: **FileNotFoundException from org.labkey.api.util.FileUtil.copyFile() 11768: **Bogus "Pipeline Override" message at root of home project? 12626: **Improve look and feel of message bar actions for data region 12714: **Participant Groups: permissions for creating shared groups should be expanded to editors

+++ api/src/org/labkey/api/reports/report/ExternalScriptEngineReport.java
@@ -193,7 +193,7 @@ public class ExternalScriptEngineReport extends ScriptEngineReport implements At
                         File src = param.getFile();
                         File dst = new File(cacheDir, src.getName());
 
-                        if (dst.createNewFile())
+                        if (src.exists() && dst.createNewFile())
                         {
                             FileUtil.copyFile(src, dst);
